chore: big refactor of lcsb.py files
