+++ core/lcsb/__init__.py
@@ -1,0 +1,5 @@
+"""
+This module contains code that is specific to LCSB's deployment and components that they use (Keycloak, REMS, IDService).
+
+The code might not be relevant in other context (and by default is disabled - you might enable it in SETTINGS.py).
+"""
